Add components.py
components.py defines an enumeration listing each type of component, a
Component class that encapsulates what a component is, and instantiates
a component for each type in the enumeration at the module level.

components.py is used by other freightplan modules whenever it needs to
know about or do anything with freighter components.