Begin tentative integration of Floors with Editor
Not sure where this will end up going. Is it reasonable to construct 441
Cell objects for every floor? If we didn't and only instantiated them
when a Tile is placed, then why have a Cell AND a Tile class? Shouldn't
Cell.setTile() handle positioning the tile, and not the Editor?