Add gui/tool.py
Defines the base class for editor tools, which provide an abstraction
for interacting with the Editor and its contents. Currently not sure if
it's better to reimplement icon, tooltip, status tip, et. al members or
just inherit/compose a QAction, given it covers all of these; I'm not
sure if this would be an appropriate application of QAction, however.