Adjust event logic in Tool
After studying the event system further, I believe this is more correct.
Embarrassingly, I got mixed up and forgot that event handlers are
supposed to set the accepted flag in lieu of returning a bool.
Tool.event has return values that are more in line with Qt's events.

What I still don't know is if Qt handles propagating the event to Tool's
parent if Tool.event() returns false, or if I have to implement that
myself. The documentation isn't clear about what Qt does for you and
what is simply part of QWidget's implementation.